model_deploy/python/requirements.txt
python-dotenv
fastapi
uvicorn
pydantic
tensorflow==2.9